requirements.txt
asgiref==3.8.1
astroid==3.3.8
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.24.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
dill==0.3.9
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
idna==3.10
isodate==0.7.2
isort==5.13.2
Markdown==3.7
mccabe==0.7.0
msal==1.31.1
msal-extensions==1.2.0
mypy==1.14.1
mypy-extensions==1.0.0
platformdirs==4.3.6
portalocker==2.10.1
psycopg==3.2.4
psycopg-binary==3.2.4
pycparser==2.22
PyJWT==2.10.1
pylint==3.3.3
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==11.5.0
tomlkit==0.13.2
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
